+++ .vortex/installer/tests/Fixtures/handler_process/hosting_acquia/docroot/modules/custom/sw_demo/src/Plugin/GeneratedContent/Node/Page.php
@@ -1,0 +1,55 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Drupal\sw_demo\Plugin\GeneratedContent\Node;
+
+use Drupal\generated_content\Attribute\GeneratedContent;
+use Drupal\generated_content\Plugin\GeneratedContent\GeneratedContentPluginBase;
+
+/**
+ * Generate page nodes.
+ *
+ * @codeCoverageIgnore
+ */
+#[GeneratedContent(
+  id: 'sw_demo_node_page',
+  entity_type: 'node',
+  bundle: 'page',
+  weight: 20,
+)]
+class Page extends GeneratedContentPluginBase {
+
+  protected const TOTAL = 20;
+
+  /**
+   * {@inheritdoc}
+   */
+  public function generate(): array {
+    $entities = [];
+
+    $storage = $this->entityTypeManager->getStorage('node');
+
+    for ($i = 1; $i <= self::TOTAL; $i++) {
+      $node = $storage->create([
+        'type' => 'page',
+        'title' => sprintf('Demo page %s %s', $i, $this->helper::staticName()),
+        'status' => 1,
+      ]);
+
+      $node->set('body', [
+        'value' => $this->helper::staticRichText(3),
+        'format' => 'full_html',
+      ]);
+
+      $node->save();
+
+      $this->helper::log('Created "%s" node "%s" [ID: %s]', $node->bundle(), $node->toLink()->toString(), $node->id());
+
+      $entities[] = $node;
+    }
+
+    return $entities;
+  }
+
+}
