--- .vortex/installer/tests/Fixtures/handler_process/hosting_acquia/docroot/modules/custom/sw_demo/src/Plugin/GeneratedContent/Taxonomy/Tags.php
@@ -1,57 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Drupal\sw_demo\Plugin\GeneratedContent\Taxonomy;
-
-use Drupal\generated_content\Attribute\GeneratedContent;
-use Drupal\generated_content\Plugin\GeneratedContent\GeneratedContentPluginBase;
-
-/**
- * Generate tags taxonomy terms.
- *
- * @codeCoverageIgnore
- */
-#[GeneratedContent(
-  id: 'sw_demo_taxonomy_term_tags',
-  entity_type: 'taxonomy_term',
-  bundle: 'tags',
-  weight: 10,
-)]
-class Tags extends GeneratedContentPluginBase {
-
-  /**
-   * Tag names to generate.
-   */
-  protected const TAGS = [
-    'Technology',
-    'Science',
-    'Health',
-    'Business',
-    'Environment',
-  ];
-
-  /**
-   * {@inheritdoc}
-   */
-  public function generate(): array {
-    $entities = [];
-
-    $storage = $this->entityTypeManager->getStorage('taxonomy_term');
-
-    foreach (self::TAGS as $name) {
-      $term = $storage->create([
-        'vid' => 'tags',
-        'name' => $name,
-      ]);
-      $term->save();
-
-      $this->helper::log('Created "%s" term "%s" [ID: %s]', $term->bundle(), $term->toLink()->toString(), $term->id());
-
-      $entities[] = $term;
-    }
-
-    return $entities;
-  }
-
-}
